common: downgrade ECCache::size_type to size_t
On ILP32 builds, size_t is 32 bits and does not match up
with the 64 bit integer returned from ECCache::MaxSize, yielding
a build error:

ECCacheManager.cpp:114:76: error: no matching function for call to
"min(size_t, long long unsigned int)"
   m_ObjectsCache.SetMaxSize(std::min(size_t(16 << 20), cell_cache_size / 16));

+++ common/ECCache.h
@@ -53,7 +53,7 @@ public:
 class _kc_export ECCacheBase {
 public:
 	typedef unsigned long		count_type;
-		typedef uint64_t	size_type;
+	typedef size_t size_type;
 
 	_kc_hidden virtual ~ECCacheBase(void) = default;
 	_kc_hidden virtual count_type ItemCount(void) const = 0;
